--- Assets/Noah-Git.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: be60e38c361ae41bd87646709a4987bb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
