--- Assets/Noah-Git/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b135a2b4bae5345469427e5e67b16702
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
