--- Assets/Noah-Git/Scripts/CameraFollow.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e897e1143657a41dba6f2ea519798373
