--- Assets/Noah-Git/Scripts/CameraShake.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e0a210cd212eb461fb8130aca126270f
