--- Assets/Noah-Git/Scripts/PlayerMovement.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 60a867c581d9c483f8fba098353631b2
