+++ Assets/Scripts/Camera/CameraFollow.cs
@@ -1,6 +1,5 @@
 using System.Collections;
 using System.Collections.Generic;
-using TMPro;
 using UnityEngine;
 
 public class CameraFollow : MonoBehaviour
