+++ Assets/Scripts/Camera/CameraFollow.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 06f54a5cc0fd97949bb46815459476ec
+guid: e897e1143657a41dba6f2ea519798373
