+++ Assets/Scripts/Camera/CameraShake.cs
@@ -10,19 +10,14 @@ public class CameraShake : MonoBehaviour
     private float magnitude;
     private bool wasShaking = false; //Used to reset after we are done
 
-    private Vector3 originalCameraOffset;
     [SerializeField]
     GameObject player;
 
-    private void Start()
+    public void ShakeCamera(float duration, float mgnitude)
     {
-        originalCameraOffset = transform.localPosition;
-    }
-
-    public void ShakeCamera(float duration, float magnitude)
-    {
-        if (enabled) {
-            StartCoroutine(Shake(duration, magnitude));
+        if (enabled)
+        {
+            StartCoroutine(Shake(duration, mgnitude));
         }
     }
 
@@ -32,6 +27,7 @@ public class CameraShake : MonoBehaviour
         //print("Camera local position " + originalPos);
 
         float elapsed = 0.0f;
+        GetComponent<CameraFollow>().enabled = false;
 
         while (elapsed < duration)
         {
@@ -48,6 +44,7 @@ public class CameraShake : MonoBehaviour
         }
 
         transform.localPosition = GetNeutralPos();
+        GetComponent<CameraFollow>().enabled = true;
     }
 
     public void setShakingTrue(float mgnitude)
@@ -56,11 +53,13 @@ public class CameraShake : MonoBehaviour
         isShaking = true;
         magnitude = mgnitude;
         wasShaking = true;
+        GetComponent<CameraFollow>().enabled = false;
     }
     public void setShakingFalse()
     {
         isShaking = false;
         magnitude = 0;
+        GetComponent<CameraFollow>().enabled = true;
     }
 
     private void Update()
@@ -81,7 +80,8 @@ public class CameraShake : MonoBehaviour
         }
     }
 
-    private Vector3 GetNeutralPos() {
-        return originalCameraOffset;
+    private Vector3 GetNeutralPos()
+    {
+        return GetComponent<CameraFollow>().getPos();
     }
 }
