+++ Assets/Scripts/Camera/CameraShake.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 8b7735f1685fd89429ff376f9b17d5a1
+guid: e0a210cd212eb461fb8130aca126270f
